Fix crash when compiled with optimization: disable long double in union Value. Seems like long double in unions is unstable.

+++ src/c.h
@@ -76,7 +76,7 @@ typedef struct table *Table;
 typedef union value {
     long i;
     unsigned long u;
-    long double d;
+    double d; // Note: long double is broken.
     void *p;
     void (*g)(void);
 } Value;
